LABIOC/requirements.txt
django >2
wmi
pypiwin32
azure
exchangelib
python-ntlm3